# class-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Class Operations API
  version: 1.5.0
  license:
    name: Proprietary
    identifier: LicenseRef-ClassOps-Proprietary
  description: |
    API cho các màn React hiện có: xác thực, lập lịch lớp và vận hành giáo viên.
    JWT có `accountScope=PLATFORM|TENANT`; tenant chỉ được suy ra cho phiên TENANT.
    API tạo lớp/ca/override không nhận link online theo DEC-046.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /teachers/me/dashboard:
    get:
      operationId: getTeacherDashboard
      tags: [Teaching operations]
      summary: KPI và các buổi hôm nay của giáo viên trong JWT
      responses:
        "200":
          description: Dashboard giáo viên
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TeacherDashboardData"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /teachers/me/classes:
    get:
      operationId: getMyTeachingClasses
      tags: [Teaching operations]
      summary: Lớp đang hoặc đã thuộc phạm vi giáo viên
      parameters:
        - {$ref: "#/components/parameters/Search"}
        - name: status
          in: query
          schema:
            type: string
            enum: ["", DRAFT, SCHEDULED, ACTIVE, AWAITING_CLOSE, CLOSED, CANCELLED]
        - name: role
          in: query
          schema: {type: string, enum: ["", PRIMARY, ACTUAL]}
        - name: from
          in: query
          schema: {type: string, format: date}
        - name: to
          in: query
          schema: {type: string, format: date}
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/PageSize"}
      responses:
        "200":
          description: Trang lớp theo phạm vi giáo viên
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TeacherClassPage"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /teachers/me/classes/{classId}/sessions:
    get:
      operationId: getMyClassSessions
      tags: [Teaching operations]
      summary: Lịch sử buổi trong nhiệm kỳ/phạm vi giáo viên
      parameters:
        - {$ref: "#/components/parameters/ClassId"}
        - name: status
          in: query
          description: Lọc trạng thái buổi; bỏ trống để lấy tất cả
          schema:
            type: string
            enum: [SCHEDULED, IN_PROGRESS, PENDING_CONFIRMATION, COMPLETED, CANCELLED]
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/PageSize"}
      responses:
        "200":
          description: Header lớp và trang buổi
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TeacherClassSessions"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /sessions/{sessionId}:
    get:
      operationId: getSessionOperationsDetail
      tags: [Teaching operations]
      summary: Chi tiết buổi, roster, hồ sơ và quyền thao tác
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
      responses:
        "200":
          description: Workspace buổi học
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionOperationsDetail"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /sessions/{sessionId}/check-ins:
    post:
      operationId: checkInTeachingSession
      tags: [Teaching operations]
      summary: Giáo viên thực tế check-in trong cửa sổ do backend tính
      description: Online bắt buộc URL http/https; tại lớp không chấp nhận URL.
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CheckInInput"}
      responses:
        "200":
          description: Trạng thái buổi sau check-in
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionOperationsDetail"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/pedagogical-record:
    patch:
      operationId: savePedagogicalRecord
      tags: [Teaching operations]
      summary: Lưu nguyên tử lesson report, điểm danh và nhận xét học sinh
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PedagogicalRecordInput"}
      responses:
        "200":
          description: Chi tiết mới nhất
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionOperationsDetail"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/tests:
    post:
      operationId: createSessionTest
      tags: [Teaching operations]
      summary: Tạo bài kiểm tra duy nhất cho buổi và khởi tạo kết quả cho roster
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SessionTestInput"}
      responses:
        "201":
          description: Chi tiết buổi sau khi tạo bài kiểm tra
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionOperationsDetail"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/tests/{testId}:
    patch:
      operationId: updateSessionTest
      tags: [Teaching operations]
      summary: Sửa bài kiểm tra và toàn bộ điểm/nhận xét học sinh trong một transaction
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/TestId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SessionTestUpdateInput"}
      responses:
        "200":
          description: Chi tiết buổi mới nhất
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionOperationsDetail"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/verification-decisions:
    post:
      operationId: decideMissingCheckInSession
      tags: [Teaching operations]
      summary: Quản lý xác nhận đã dạy hoặc hủy buổi thiếu check-in
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VerificationDecisionInput"}
      responses:
        "200":
          description: Buổi Completed hoặc Cancelled
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionOperationsDetail"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ClassId: {name: classId, in: path, required: true, schema: {type: string, format: uuid}}
    SessionId: {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
...
    TestId: {name: testId, in: path, required: true, schema: {type: string, format: uuid}}
    Search: {name: search, in: query, schema: {type: string, default: ""}}
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 150}
...
  responses:
    ValidationError:
      description: Dữ liệu không hợp lệ
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    Forbidden:
      description: Không có quyền
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    NotFound:
      description: Không tìm thấy tài nguyên trong tenant
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Conflict:
      description: |
        Preview stale, TEACHER_OVERLAP, ROOM_OVERLAP, STUDENT_OVERLAP,
        optimistic lock hoặc idempotency conflict.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
  schemas:
...
    SessionAction:
      type: string
      enum: [SUBSTITUTE_TEACHER, CANCEL_SESSION, CREATE_MAKEUP]
...
    TeacherDashboardData:
      type: object
      required: [date, teacherName, metrics, sessions]
      properties:
        date: {type: string, format: date}
        teacherName: {type: string}
        metrics:
          type: object
          required:
            [todaySessions, checkInAvailable, missingDocumentation, monthTeachingHours, monthAccruedSalary]
          properties:
            todaySessions: {type: integer, minimum: 0}
            checkInAvailable: {type: integer, minimum: 0}
            missingDocumentation: {type: integer, minimum: 0}
            monthTeachingHours: {type: number, minimum: 0}
            monthAccruedSalary: {type: number, minimum: 0}
        sessions:
          type: array
          items: {$ref: "#/components/schemas/TodayTeachingSession"}
    TodayTeachingSession:
      type: object
      required:
        [id, classId, classCode, className, ordinal, startAt, endAt, mode, status,
         checkInState, checkInOpensAt, substitution]
      properties:
        id: {type: string, format: uuid}
        classId: {type: string, format: uuid}
        classCode: {type: string}
        className: {type: string}
        ordinal: {type: integer, minimum: 1}
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomName: {type: [string, "null"]}
        status: {$ref: "#/components/schemas/SessionStatus"}
        checkInState: {$ref: "#/components/schemas/CheckInState"}
        checkInOpensAt: {type: string, format: date-time}
        substitution: {type: boolean}
    TeacherClassItem:
      type: object
      required:
        [id, code, name, status, teacherRole, startDate, completedSessions, totalSessions]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        status: {type: string}
        teacherRole: {type: string, enum: [PRIMARY, ACTUAL]}
        startDate: {type: string, format: date}
        expectedEndDate: {type: [string, "null"], format: date}
        completedSessions: {type: integer, minimum: 0}
        totalSessions: {type: integer, minimum: 1}
        latestSessionAt: {type: [string, "null"], format: date-time}
    TeacherClassPage:
      allOf:
        - {$ref: "#/components/schemas/Page"}
        - type: object
          properties:
            items:
              type: array
              items: {$ref: "#/components/schemas/TeacherClassItem"}
    TeacherClassSessions:
      type: object
      required: [learningClass, sessions]
      properties:
        learningClass:
          type: object
          required: [id, code, name, status, completedSessions, totalSessions]
          properties:
            id: {type: string, format: uuid}
            code: {type: string}
            name: {type: string}
            status: {type: string}
            completedSessions: {type: integer}
            totalSessions: {type: integer}
        sessions:
          allOf:
            - {$ref: "#/components/schemas/Page"}
            - type: object
              properties:
                items:
                  type: array
                  items: {$ref: "#/components/schemas/TeacherSessionSummary"}
    TeacherSessionSummary:
      type: object
      required:
        [id, ordinal, startAt, endAt, status, actualTeacherName, actualTeacher, readOnly,
         lessonName, participatedStudents, rosterStudents, missingDocumentation, hasTest,
         testResultCount]
      properties:
        id: {type: string, format: uuid}
        ordinal: {type: integer}
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        status: {$ref: "#/components/schemas/SessionStatus"}
        actualTeacherName: {type: string}
        actualTeacher: {type: boolean}
        readOnly: {type: boolean}
        lessonName: {type: string}
        participatedStudents: {type: integer}
        rosterStudents: {type: integer}
        missingDocumentation: {type: boolean}
        hasTest: {type: boolean}
        testResultCount: {type: integer}
    SessionStatus:
      type: string
      enum: [SCHEDULED, IN_PROGRESS, COMPLETED, PENDING_CONFIRMATION, CANCELLED]
    CheckInState:
      type: string
      enum: [TOO_EARLY, OPEN, CHECKED_IN, WINDOW_CLOSED, COMPLETED, CANCELLED]
    AttendanceStatus:
      type: string
      enum: [PRESENT, LATE, LEFT_EARLY, ABSENT_EXCUSED, ABSENT_UNEXCUSED]
    SessionTest:
      type: object
      required: [id, testName, maxScore, testDate, comment, version]
      properties:
        id: {type: string, format: uuid}
        testName: {type: string}
        maxScore: {type: number, exclusiveMinimum: 0}
        testDate: {type: string, format: date}
        comment: {type: string}
        version: {type: integer, format: int64}
    TestResult:
      type: object
      required: [id, comment, version]
      properties:
        id: {type: string, format: uuid}
        score: {type: [number, "null"], minimum: 0}
        comment: {type: string}
        version: {type: integer, format: int64}
    RosterStudent:
      type: object
      required:
        [studentId, code, name, attendanceNote, attendanceVersion, sessionComment,
         commentVersion, testResult]
      properties:
        studentId: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        attendanceStatus:
          oneOf:
            - {$ref: "#/components/schemas/AttendanceStatus"}
            - type: "null"
        attendanceNote: {type: string}
        attendanceVersion: {type: integer, format: int64}
        sessionComment: {type: string}
        commentVersion: {type: integer, format: int64}
        testResult:
          oneOf:
            - {$ref: "#/components/schemas/TestResult"}
            - type: "null"
    SessionOperationsDetail:
      type: object
      required:
        [id, classId, classCode, className, ordinal, startAt, endAt, mode, status,
         plannedTeacherId, actualTeacherId, actualTeacherName, substitution, makeup,
         allowedActions, actualTeacher, canEdit, canVerify, checkInState, checkInOpensAt,
         rosterFrozen, rosterRevision, missingDocumentation, version, lessonReport, sessionTest, students,
         participatedStudents]
      properties:
        id: {type: string, format: uuid}
        classId: {type: string, format: uuid}
        classCode: {type: string}
        className: {type: string}
        ordinal: {type: integer}
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomId: {type: [string, "null"], format: uuid}
        roomName: {type: [string, "null"]}
        onlineLink: {type: [string, "null"], format: uri}
        status: {$ref: "#/components/schemas/SessionStatus"}
        plannedTeacherId: {type: string, format: uuid}
        actualTeacherId: {type: string, format: uuid}
        actualTeacherName: {type: string}
        substitution: {type: boolean}
        makeup: {type: boolean}
        makeupRootSessionId: {type: [string, "null"], format: uuid}
        replacesSessionId: {type: [string, "null"], format: uuid}
        replacementSessionId: {type: [string, "null"], format: uuid}
        cancellationReason: {type: [string, "null"]}
        allowedActions: {type: array, items: {$ref: "#/components/schemas/SessionAction"}}
        actualTeacher: {type: boolean}
        canEdit: {type: boolean}
        canVerify: {type: boolean}
        checkInState: {$ref: "#/components/schemas/CheckInState"}
        checkInOpensAt: {type: string, format: date-time}
        checkIn:
          type: [object, "null"]
          properties:
            checkedInAt: {type: string, format: date-time}
            onlineLink: {type: [string, "null"], format: uri}
        rosterFrozen: {type: boolean}
        rosterRevision: {type: string}
        missingDocumentation: {type: boolean}
        version: {type: integer, format: int64}
        lessonReport:
          type: object
          required: [lessonName, lessonContent, version]
          properties:
            lessonName: {type: string}
            lessonContent: {type: string}
            recordUrl: {type: [string, "null"], format: uri}
            version: {type: integer, format: int64}
        sessionTest:
          oneOf:
            - {$ref: "#/components/schemas/SessionTest"}
            - type: "null"
        students:
          type: array
          items: {$ref: "#/components/schemas/RosterStudent"}
        participatedStudents: {type: integer, minimum: 0}
    CheckInInput:
      type: object
      required: [version]
      properties:
        onlineLink: {type: [string, "null"], format: uri}
        version: {type: integer, format: int64, minimum: 0}
    PedagogicalRecordInput:
      type: object
      required: [rosterRevision, lessonReport, students]
      properties:
        rosterRevision: {type: string}
        lessonReport:
          type: object
          required: [lessonName, lessonContent, version]
          properties:
            lessonName: {type: string}
            lessonContent: {type: string}
            recordUrl: {type: [string, "null"], format: uri}
            version: {type: integer, format: int64, minimum: 0}
        students:
          type: array
          items:
            type: object
            required:
              [studentId, attendanceNote, attendanceVersion, sessionComment, commentVersion]
            properties:
              studentId: {type: string, format: uuid}
              attendanceStatus:
                oneOf:
                  - {$ref: "#/components/schemas/AttendanceStatus"}
                  - type: "null"
              attendanceNote: {type: string}
              attendanceVersion: {type: integer, format: int64, minimum: 0}
              sessionComment: {type: string}
              commentVersion: {type: integer, format: int64, minimum: 0}
    SessionTestInput:
      type: object
      required: [testName, maxScore, testDate]
      properties:
        testName: {type: string, minLength: 1}
        maxScore: {type: number, exclusiveMinimum: 0}
        testDate: {type: string, format: date}
        comment: {type: string}
    SessionTestUpdateInput:
      allOf:
        - {$ref: "#/components/schemas/SessionTestInput"}
        - type: object
          required: [version, rosterRevision, results]
          properties:
            version: {type: integer, format: int64, minimum: 0}
            rosterRevision: {type: string, minLength: 1}
            results:
              type: array
              items:
                type: object
                required: [studentId, comment, version]
                properties:
                  studentId: {type: string, format: uuid}
                  score: {type: [number, "null"], minimum: 0}
                  comment: {type: string}
                  version: {type: integer, format: int64, minimum: 0}
    VerificationDecisionInput:
      type: object
      required: [decision, reason, version]
      properties:
        decision: {type: string, enum: [CONFIRM_TAUGHT, CANCEL]}
        reason: {type: string, minLength: 1}
        version: {type: integer, format: int64, minimum: 0}
...
    Page:
      type: object
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items: {type: array, items: {}}
        page: {type: integer, minimum: 1}
        pageSize: {type: integer, minimum: 1}
        totalItems: {type: integer, format: int64}
        totalPages: {type: integer}
    Problem:
      type: object
      required: [type, title, status, detail, code, message, timestamp, traceId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code: {type: string}
        message: {type: string}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        details: {type: object, additionalProperties: true}
        retryAfterSeconds: {type: integer, minimum: 1}
        timestamp: {type: string, format: date-time}
        traceId: {type: string}